Create DeGraciaMedina_HW1.md
Name: Xenia De Gracia Medina.   Date: August 31, 2020.     Homework #1

The way I made my forecast was by choosing the last 10 years of information and making an average with the exact week of all those 10 years. This method was used just for Week 1 and 2.

For the entire semester forecast, I used just the data of the year 2019 to predict the same conditions for this current year.